Creacion de la clase multiplicador exclusividad|
Esta clase sirve para diferenciarse del multiplicador comun
