+++ src/main/java/edu/fiuba/algo3/modelo/preguntas/resultados/EstadoCorrectoClasico.java
@@ -1,9 +1,6 @@
 package edu.fiuba.algo3.modelo.preguntas.resultados;
 
-import edu.fiuba.algo3.modelo.preguntas.modificadores.AnalizadorExclusividad;
-import edu.fiuba.algo3.modelo.preguntas.modificadores.Multiplicador;
-import edu.fiuba.algo3.modelo.preguntas.modificadores.UsuarioRespondioBien;
-import edu.fiuba.algo3.modelo.preguntas.modificadores.UsuarioSeEquivoco;
+import edu.fiuba.algo3.modelo.preguntas.modificadores.*;
 import edu.fiuba.algo3.modelo.preguntas.puntos.Punto;
 import edu.fiuba.algo3.modelo.preguntas.puntos.PuntoNulo;
 import edu.fiuba.algo3.modelo.preguntas.puntos.PuntoPositivo;
