+++ src/main/java/edu/fiuba/algo3/modelo/preguntas/resultados/Resultado.java
@@ -3,6 +3,7 @@ package edu.fiuba.algo3.modelo.preguntas.resultados;
 
 import edu.fiuba.algo3.modelo.preguntas.modificadores.AnalizadorExclusividad;
 import edu.fiuba.algo3.modelo.preguntas.modificadores.Multiplicador;
+import edu.fiuba.algo3.modelo.preguntas.modificadores.MultiplicadorJugador;
 import edu.fiuba.algo3.modelo.preguntas.puntos.Punto;
 
 public interface Resultado {
