+++ src/main/java/edu/fiuba/algo3/modelo/preguntas/resultados/ResultadoPenalizable.java
@@ -2,6 +2,7 @@ package edu.fiuba.algo3.modelo.preguntas.resultados;
 
 import edu.fiuba.algo3.modelo.preguntas.modificadores.AnalizadorExclusividad;
 import edu.fiuba.algo3.modelo.preguntas.modificadores.Multiplicador;
+import edu.fiuba.algo3.modelo.preguntas.modificadores.MultiplicadorJugador;
 import edu.fiuba.algo3.modelo.preguntas.puntos.Punto;
 import edu.fiuba.algo3.modelo.preguntas.puntos.Puntuacion;
 import edu.fiuba.algo3.modelo.preguntas.puntos.PuntoNegativo;
